Extracts HtmlEditorKit, removes keep font attribute, not needed.
