+++ idea-plugin/src/main/kotlin/arrow/meta/ide/plugins/initial/tooltip/formatting/Styles.kt
@@ -1,3 +1,3 @@
-package arrow.meta.ide.plugins.initial.tooltip
+package arrow.meta.ide.plugins.initial.tooltip.formatting
 
 const val cdnStyle = "https://47deg-academy.s3.amazonaws.com/css/edutools-style-light.css"
